configuring project structure
