+++ client/src/main.jsx
@@ -2,14 +2,10 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
 import App from './App.jsx'
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 
-const queryClient = new QueryClient()
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
-    <QueryClientProvider client={queryClient}> {/* обязательная обертка для работы query-запросов (API) */}
       <App />
-    </QueryClientProvider>
   </StrictMode>,
 )
